check_saturation updated
check_saturation now functions as expected, and can be adjust to just view a subset of rows (to avoid snow glare which will saturate areas we aren't interested in)

Very basic start to gui acquisition widgets made

+++ pycam/setupclasses.py
@@ -124,7 +124,7 @@ class CameraSpecs:
         self.min_saturation = 0.7   # Minimum saturation accepted before adjusting shutter speed (if auto_ss is True)
         self.max_saturation = 0.9   # Maximum saturation accepted before adjusting shutter speed (if auto_ss is True)
         self.saturation_pixels = 100  # Number of pixels checked for saturation
-        self.saturation_rows = self.pix_num_y / 2   # rows to extract for saturation check (don't want to check lower rows as snow may be present
+        self.saturation_rows = int(self.pix_num_y / 2)   # rows to extract for saturation check (don't want to check lower rows as snow may be present
 
     @property
     def bit_depth(self):
